+++ src/main/scala/com/diatom/agent/deletoragent.scala
