+++ src/main/scala/com/diatom/agent/fitness.scala
