--- src/main/scala/com/diatom/island/TTerminationCriteria.scala
@@ -1,9 +1,0 @@
-package com.diatom.island
-
-import scala.concurrent.duration.Duration
-
-trait TTerminationCriteria {
-  def time: Duration
-}
-
-case class TerminationCriteria(time: Duration) extends TTerminationCriteria
