+++ src/main/scala/com/diatom/paretofrontier.scala
