+++ src/main/scala/com/diatom/scored.scala
